requirements.txt
redis
-e git+https://github.com/dotcloud/docker-py.git#egg=docker
pyyaml
jsonpath-rw
